fix: load config directly into db
Importing Infra configs via API adds an unnecessary dependency on the
API server. This adds a layer of authn/authz that isn't necessary for a
local import. If an user decides to disable the admin access key or
installs on after the fact through `/v1/setup`, import will fail.

Instead, load the data directly into the database. As before, if an
identity or group for a grant does not exist, the import process will
create a placeholder, provided the provider source is known or can be
inferred.

On server start, configs will be loaded from file and any extraneous
providers or grants will be removed. A future enhancement, #1118, can be
implemented to fine tune this behaviour to only remove resources the
config loader created.

By loading the config directly, we can also, at least temporarily remove
the need for an `import` CLI command. While this will definitely be
useful in the long run, it is currently unnecessary.