fix(engine): remove Recovery middleware
gin.Recovery middleware is used to recover from panics which is new
behaviour added during the gin move. Previously, panics were not
recovered.

This new middleware produces the recovery message seen when requests
are aborted, which is common with requests with `watch=true` or, using
kubectl, `kubectl get pods -w` or log following. Aborts in this case
should not be considered an error.