style(cli): remove some short flags from CLI commands
Short flags are nice for convenience, but once we add them it is almost
impossible to remove them. I think if we start with fewer short flags it
will be easier to add them in the future.

The risk of adding them all now is that we won't be able to use them for
other CLI flags in the future, and the short flag for one command may
be different from the short for a different command. For example the -f
for logout is much different than the -f for server or connector.

BREAKING CHANGE: some CLI short flags were removed.