fix: generation of openapi spec when there are no changes
Using reflect.Equal does not appear to work, there must be differences
between how the spec is decoded from a file, and generated from types.

This commit goes back to comparing the encoded version of the spec. That
works better, but requires that we don't append servers, otherwise the
old spec will always have different servers.

All of the other fields we set are from writeOpenAPISpec replace any
existing values, so it seems appropriate to do the same for Servers.