fix: http.Transport not using reasonable defaults
HTTP clients in the connector are not using the reasonable defaults set
in net/http. This leads to excessive amounts of open connections being
kept open and resource leakage. In the long term, this issue will cause
the connector to no being able to query the server which means the
connector will become out of sync. This lead to resource leakage in the
server since each connection needs to be kept open.

This may also be the cause of the memory leak as mentioned in
https://github.com/infrahq/infra/issues/1361

Use the default transport as defined in net/http. Use a single transport
for connector <> server communication and another transport for
connector <> Kubernetes API server.