fix: use a more human readable format for logs from the CLI
Previously we were using the same logger for both long running processes
(server and connector) and short lived ones (other CLI commands).

The logger did use a text format (non-json) when stdin is a terminal,
but the log would contain a long timestamp and function name+line
number, which is probably not what users want to see from a CLI.

This commit keeps the existing logger for long running processes, but
uses a less verbose format for CLI commands. The new format is always
text, and removes the timestamp and function name + line number.

Also changed the logging of error from the CLI to use the logger so that
output of warnings and errors is more consistent.

Also removed some unnecessary error return values, and removed some
very short functions to make the code easier to trace.