refactor: call gin.SetMode in fewer places
I noticed we were calling gen.SetMode in few places. Sometimes in init()
sometimes right before creating the router, and sometimes in tests.

When I was adding new routes I found it helpful to be able to turn on
debug mode with an environment variable, but all of these calls to
gin.SetMode made it difficult.

Also looking over the code, and seeing github.com/gin-gonic/gin/issues/2984,
it seems that TestMode does nothing. It is effectively equivalent to
ReleaseMode. Only DebugMode changes what is logged.

The new behaviour works as follows:
1. `gin` itself defaults to DebugMode in `mode.go:init()`
2. the two packages that use gin (server and connector) both call
   `ginutil.SetMode` before starting a gin router.
3. `ginuutil.SetMode` will attempt to use the same environment variable
   as `gin` would have used, but instead defaults to ReleaseMode when
   not set.

I believe this gives us the correct behaviour. We should always be in
ReleaseMode. Developer debugging locally and operators running a binary can set
`GIN_MODE=debug` if they need to.

I could not find an appropriate place to put this new function, so I had
to create the `ginutil` package. I'm happy to move it somewhere else if anyone
has suggestions.