improve(ci): remove from_name from labels.yaml
This field was used to migrate existing labels. Now that the migration
is done the field can be removed.

Also adds one more label that I missed on the first pass, `area/secrets`.

Also changes all the color values to lowercase. It seems that github
stores these as lowercase, so changing them to lowercase makes the
preview nicer (shows it as skipping) and will avoid doing unnecessary
updates to labels.