improve(cli): move newServerCmd to its own file
This changes follows the convention used by most Go CLIs of putting the command in a file
named after the command. It will make it easier for anyone reading the code to quickly find
the server command.

It is being done now in preparation for adding defaultOptions to this file.