fix(api): scope api middleware to api routes
Use a gin.RouterGroup in API.registerRoutes so that the routes only apply to
the API handlers.

I attempted to use a gin.RouterGroup for the UI routes, but it does not
work because the UI and UI proxy work on all paths. There is no handler
for the UI or separate prefix we can use to add handlers.

Also moved the middleware setup into GenerateRoutes to remove the need
to add them separtely for the UI.

Co-Authored-By: Daniel Nephin <dnephin@infrahq.com>