improve(metrics): restore standard prometheus metrics
These two collectors are normally added to DefaultRegistry from an init
function. I noticed while manually testing that we no longer had these.

The Middleware function is a bit of an awkward place to register them,
but it's probably fine for now since both the server and connector use
this middleware.

Also improves the test to show that at least one metric from each of
these collectors is in the output.