improve: log identity name for logged in API request
Removes the IdentityAwareMiddleware as it did not work, and because we
can do this with a single middleware. It did not work for a few reasons.
It attempted to read an identity from the context before it existed, and
because the identity in the context is a model.Identity, not just an ID.

This new approach works because we read the identity out after the
request handler runs.

Tested manually and saw authenticated requests with an identity field,
and unautheticated requests without.