improve(logging): remove logging helpers
These helpers for retrieving a logging from a context were not widely
used.

Receiving a logger from a context is a nice approach, but until we
unexport logging.L and logging.S it introduces inconsistency, because
there are multiple ways of getting a logger.

For now let's have everything use logging.L or logging.S. In the future
we can look at passing the logger exclusively through context.Context.

Also use less sugar.