improve: structure the UI config
A structured config allows us to pass around parts of the config to the
components that use them. It makes code easier to trace.

The `enableUI` field is now `ui.enabled` and the `uiProxyURL` field is
now `ui.proxyURL`.

BREAKING CHANGE: the UI fields in the server config file have changed.