fix(cli): fix the extension-deadline flag
I started by changing the fields from string to time.Duration to avoid
the need to ParseDuration. Then I wrote a test for the command.

The test was failing because --extension-deadline was not being passed
to the API. I'm not sure why, the struct field tags look right.

I changed the command to use DurationVar, which got the test to pass.

This change removes support for setting these fields from environment
viarables, but that should be an easy thing to restore. I am working on
a proposal for how we can do that with a single function, without the
need for viper.

Since this is a CLI command that we expect humans to run, I guess no one
is trying to use it with environment variables right now.