improve: add a test for CreateGrant API endpoint
This is the first test that is strict about the API response body, so it
includes some new helper functions.

As we use these more, we may want to extract them into an apitest
package.

There are a few fields here that are not predictable (uid and dates),
which makes writing a strict test more difficult.

The approach taken here is to do an approximate comparison. An
alternative would be to patch the uid generator and the database
time.Now to set specific values. I think the approximate comparison
might be easier to maintain than a lot of patching.