fix: cli log level and reading from env vars
spf13/cobra only runs a single Command.PersistentPreRun function. If
there are multiple parent commands, and they specific more than one
PersistentPreRun, then only the last one is run.

The PersistentPreRun functions defined in grouping commands was
clobbering the one set on the RootCmd, where we set log level and were
previously calling parseOptions and are now calling DefaultsFromEnvVars.

This commit fixes the problem by moving the logic to a function and
calling it from all the PersistentPreRun functions.

We could also solve this by never using PersistentPreRun, and instead
have each command.Run call the necessary things. I think that might be a
good approach as well, but this seemed like a smaller change for now.