improve(cli): handling of non-interactive option
Instead of checking if stdin is a terminal at all call sites, use
"stdin is a terminal" as the default value for the non-interactive flag.

This approach has two advantages:
1. it allows someone to force interactive mode with
   --non-interactive=false
2. the field on the option struct is now correct and can be used
   directly, without having to remember to call other functions.

As a consequence of this change, this fixes 'infra id edit' to properly
default to the right interactivity mode. Previously it was only using
the command line flag.