improve: add tests to catch default flag value
For long-running commands (infra server, and infra connector) all
options can be set from the config file, and most can be set from env vars,
but only a few will be set from command line flags.

Command line flags accept default values, but anything that uses
cliopts.Load can set defaults on the struct before calling Load.

Because of these two constraints, we end up with two options:

1. have cliopts honor the default values set in flags, allowing defaults
   to be specified in two different ways.
2. only allow defaults to be set on the struct before calling
   cliopts.Load.

This commit goes with the second option, ignore defaults set on flags so
that we only have one way of specifying default values.

To prevent surprises, this commit also adds two tests for the flags on
these commands. The tests check that the flags all use the zero value as
the default. This should point the contributor to the right place if
they accidentally add a non-zero value on a flag.