maintain: fix a flaky test
The test was flaky because of the ordering. It was using the admin
accessKey and relying on a previous test case running first to set the
lastSeenAt.

This commit fixes the problem by using an access key for the identity
being queries. So the request in the test case sets lastSeenAt.