maintain: fix a potentially flaky test
The metrics server subtest was looking for the http_request_duration_seconds metric
which only exists after an API request is performed. This was relying on other tests having
run first.

This commit fixes the problem by using runStep, so that each subtest has to run in order,
and adding a call to /v1/version to make sure at least one API request was handled.