maintain: add some guardrails to ci
If a test modifies a file in git, or leaves a uncommited file then every
developer running the tests will need to clean up those files.

This CI check will inform any contributor that their PR leaves stale
files, which should prevent a commit from being merged to main if it leaves
the git checkout in a dirty state.