maintain: remove package-level openapi document
Package-level variables make testing more challenging and can make code
harder to reason about. They introduce side-effects to functions that
can often be avoided.

This commit replaces the boolean to skip generation in tests (which
likely makes the tests less like production) and replaces it with an
openapi3.T document on the API.

The document is now returned from GeneratesRoutes, which also removes
the risk of something referencing the statis variable before it is fully
populated.