improve: change OpenAPI functions to accept a schema
It looked a bit strange previously because we were passing a struct (not a pointer)
that was being updated.
It worked because the field being updated was a map, which is always pass by reference.
This commit changes the args to pass only the mutable map, to prevent surprises in the
future if someone tried to modify some other part of the struct.