improve: use a longer timeout in CI
This test passes reliably for me locally with only 110ms timeout, but fails sometimes in CI
with a timeout of 2 seconds.

To address this, use a much longer timeout in CI, but keep a shorter timeout for local dev
to prevent waiting a long time for the test to fail.