improve: use types to describe openapi schema
This commit moves schema descriptions from being hardcoded in the
openapi doc generation, to being methods on the types being described.

This method is the same interface we use in internal/valiate.

By using an interface it makes it easy for us to support new api types
without modifying the openapi doc generation.