fix: destination CA field encoding
encoding/json encodes []byte as a base64 encoded string. We were using
[]byte for the destination.connection.CA field because we want to
preserve the trailing newline in that field.

Instead of using a []byte, we can use a custom type.

This commit changes our generic trimWhitespace to only operate on string
types (not custom types), allowing us to use a custom type to indicate
that the value should not be normalized in the same way.