improve: close HTTP servers
This shouldn't impact production at all, but when running a server for tests we were not
releasing the ports until the process exited. When running with -count=100 this could
consume all ports.

With this change the server is closed before the test ends, which closes the listener
and releases the ports.