maintain: use a longer timeout with -race
When the login test times out the t.Cleanup function could attempt to
close the pty at the same time as the login command running in a
separate goroutine is attempting to read from it.

We can't fix the data race because the survey library does not provide
any way to cancel the prompt. So instead we have to try and prevent
the timeout.

This PR increases the timeout used with expect when the race detector is
enabled.