maintain: move openapi3 spec to testdata
Previously we were writing the openapi3 doc to the path that would be used for
public docs.
This caused us problems because after a release we had to update the doc
to the new version.

This commit moves the openspi3.json doc testdata. This allows us to
preview any API changes, and catch any problems with doc generation in
any PR that makes changes, without the problem of having to update the
version.

As part of releasing the website we will generate this doc again, so the
version will be correct on the website.