fix: use SNI in the CLI
We attempt to generate a certificate for the request based on the SNI
name or the local address. The local address lookup doesn't work when
the server is behind a load balancer.

This commit has the CLI use an SNI name so that the server can generate
a certificate with the correct IP or hostname.