maintain: remove some wrapping around zerolog
We should strive to only have one way to do things. By having multiple
ways to log, it becomes harder to audit, and harder for contributors to
find the right way.

We had to introduce some convenience methods in the logging package when
switching logging libraries, but we should limit our use of these and
slowly convert everything to a single interface.

If we want that interface to be our own, we'll need some way of exposing
the zerolog context, so that we can add structured fields to the log.

For now, this commit removes two of the unused ones, and replaces the
calls to panic.

Fatalf should never be used, so we should not have a second way to call
it.

Panicf probably should not be used, but for now at least improve the log
message by including an error field.