fix: create a separate session for each db query
Previously we were creating a new session for each request with
db.WithContext in the DatabaseMiddleware, but queries within a single
request could accidentally mutate the session for other queries in the
same request.

With this change we should be creating a new session before modifying
gorm.DB, which should prevent one query from interfering with others.