maintain: remove static settings variable
Use the settings returned by InitializeSettings, and store it on the Telemetry struct
instead of a package-level static variable.
Remove the error return from NewTelemetry, a nil db is a programming
error that should panic.
Remove configureTelemetry. Call NewTelemetry directly and assign it to the server field
directly, makes the the code easier to trace.

Call Telemetry.Close to flush telemetry, previously this was not called
anywhere.

Remove unused SaveSettings.

Add tests for InitializeSettings.