improve: never sample error requests
We are not yet using the gin.Context.Errors to track request errors, so change the logging
middleware to check the response code instead of the context errors.

Also disable sampling when the log level is set to debug or trace. In that case the user is
attempting to get all the data, so we don't want to sample.

Remove changing the log level based on error for now. That can be restored in the future when
sendAPIError does the right thing.