fix: force release version in goreleaser
Unfortunately goreleaser's release version is firmly tied to git tags.
In this case, this coupling is detrimental as there's an intermediate
step which requires goreleaser to build artifacts for a nonexistent tag.

Goreleaser doesn't have many options to modifying its version string so
we have to resort to always building a snapshot which derives its value
from the env var RELEASE_NAME. This also means any job that wants to run
the build workflow must set RELEASE_NAME

https://goreleaser.com/customization/snapshots/