maintain: fix test failure
This test depends on the product version, so the expected value changes
every time we create a new release branch, and after every release.

This commit fixes the problem by using the existing shim we added for
openapi doc generation. We can patch that shim in a test to set a
hardcoded version.

This way the expected value does not change when the product version
changes.