improve: reduce the number of places we look up values from in-cluster
Previously we would only look for in-cluster config. In the future we
should be able to specify these values from config options, so that we
can run a connector in tests.

This change reduces the number of places we use the in-cluser calls by
relying on the rest.Config that is populated by the NewKubernetes
constructor.

Also read the CA cert from the existing config, instead of reading the
file again.

Also remove the SecretReader field, it was never used.

Also add more godoc.