fix: remove direct dependency on aws-sdk-go
Use the AWS API directly to get AWS cluster name. Use a combination of
instance metadata and AWS API. The instance metadata tags is not
sufficient because 1) it requires the user to explicitly enable
exporting instance tags (it's not enabled by default) and 2) the format
of the tags percludes it from being exposed through instance metadata
(they contain `/`).

Unfortunately, cannot remove aws-sdk-go completely because
infrahq/secrets depends on it