improve: remove msg for api errors
`Msg` and similar functions sets the `message` field in the log json.
It's not very useful in the case of API errors because the interesting
bits are in the other attributes. Having a message field causes the
`message` to be highlighted in aggregators like Datadog which makes it
harder to see the interesting bits.