improve: add code generation for table methods
Also add a test that can be run with a command line flag to generate
methods of some types. A test is used, instead of a separate script,
because it makes it easy import the migratable list, which is needed
to reflect the type.

Note: it's important to re-use the fileset with the template so that
positional information about the ast is preserved. Also adding newlines
as Doc comments helps.