improve: remove second call to gorm.AutoMigrate
From now on all migrations must be explicit, no more auto-creating of columns or
constraints using AutoMigrate.

The migration test is updated to dump both the intial schema and the
migrated schema, to ensure that they match.

Removing this extra call to AutoMigrate required making a few changes:

Add migrations that were previously done implicitly by the calls to AutoMigrate.

For some reason the schema file we use for migrations has composite
primary keys and columns in a different order from the initial schema.

I guess gorm.AutoMigrate changes the order based on something in the
code. I attempted to re-order the list of tables in `initializeSchema`
many times, but nothing changed the order.

So for now I made some small changes to the saved schema which allows
the test to pass.

Also re-order the list in initializeSchema to have it match the migrated
schema.

Also re-order some fields in the models struct, so that they match the
order in the migrated schema.