maintain: restore sqlite support for migration test
Until we are totally ready to remove sqlite support for production use cases.

This commit removes the CREATE INDEX, which will continue to be created by gorm.AutoMigrate. This was necessary
because sqlite doesn't support 'USING btree'

Also removed the timezones from some fixture data, to allow it to work with sqlite.