improve: fix infinite recursion in test
I'm not sure why switching to testing with postgres caused this problem.
I suspect it is because the sqlite driver does not use any values from context,
where as the postgres driver does look for a value in the context.

Before this change the test was crashing (fatal error, not panic) because of

  runtime: goroutine stack exceeds 1000000000-byte limit

There was an infinite loop looking up a value in the context.
With this change the test passes.