improve: remove gorm from migrator
Add a shim that allows it to continue to test using the gorm DB. When we
change the data package to create a postgres connection directly we can
change the migrator as well.

Also remove transactions from migrator

Transactions can be managed by the caller. This will make it easier to switch away from gorm.