improve: move panic to a better place
If the model being saved to the database already has an OrgID set on it
then we should not need to panic. This is common when we've already
queried the database for the object we are about to modify.

This allows us to remove one of the places where we unnecessary update
a data.Transaction.