improve: remove misleading calls to new transaction
The data.NewTransaction construct is a bit misleading because at least
for now that does not actually start a transaction. Instead wraps a
gorm.DB with our transaction wrapper.

This PR makes that less confusing by removing calls to NewTransaction
and replaces them with WithOrgID. These calls should make it much more
obvious what is being done.

In the near future we are going to use the stdlib for transactions. At
that time we will likely move transaction creation into the data package
and we'll hopefully be able to remove the NewTransaction constructor

Also remove a couple places we were creating a RequestContext
unnecessarily in tests.