maintain: move an access test to be an api test
The tests written at the access layer make our code harder to change because they require
duplicating all of the work done by our middleware.

Moving this test also gives us better test coverage , because the handler and API response are
also covered by the test.

This is being done to reduce the number of places we need to call data.NewTransaction.