maintain: remove string function
Functions with more than 3 arguments are hard to use, especially when some of the args
have the same type. From the caller it's not easy to see what the values represent.

Using a string makes the code more obvious because we can see the value 10 is for
length. The field names help document what the values mean.

Also remove a TODO, the password critiera are validated in the access package.