maintain: move logging middleware to the server package
Primarily this is being done so that we can add the userID and orgID to
the request. To fetch the user ID from the request we need to import the
access package, but the access package imports the logging package. So
attempting to get the userID caused an import cycle.

Since this middleware imports gin.Context, it's not really appropriate
for a logging package that gets imported by almost every package. This
middleware is specific to the API server, so we can keep it in the server
package.