improve: add userID and orgID to api request log
Also add a message "API request completed" so that users can find these
log messages with an easy to remember search string.

Also move the fields around a bit so that we add all the fields in the
same place, instead of creating a logger, then an event, and feeding it
back into the logger.