improve: db connection string
Previously the db connection string could either be entirely set by the
dbConnectionString option, or the user had to set dbHost if they wanted
to use the other options.

With this change the user can set dbConnectionString for most options,
and use only dbPassword to pull the secret value from a safe place.

This should allow us to deprecate most of the other db options at some
point.