maintain: remove requires update method
The TODOs here suggested that we might be able to remove this field entirely, but it looks
like the associated issue was closed and we may still need it. I suspect we should keep this
field instead of trying to infer it from some other data, so that we can use it in other cases
in the future.

For example, we may have to invalidate passwords and force someone to create a new one, and this
field would work great for that purpose.

The RequiresUpdate method is removed in favor of returning the value on the existing struct. This
also removes the need to query for the user and credentials a second time.