improve: expose method to get the stdlib sql DB
In the future this should return the value of a field on data.DB. For now we can get the reference
from gorm. The panic is added instead of an error because if we can't get the sqlDB this represents a
programming error. We always call this on the data.DB, which should always have the right state.

This commit removes another call to GormTxn.