improve: make linter a separate module
The linter is run by golangci-lint as a go plugin. One of the limitations of go plugins is that
they have to be compiled with the exact same version of Go, AND the exact same version of all
go modules.

The linter uses many of the same modules from golang.org/x, so we frequently run into conflicts
when we update versions of the modules we use.

To address this problem we can make the linter a separate module. This module can keep its
dependencies aligned with the version used by golangci-lint, without having to keep older
versions of dependencies in the main infra module.