feat: connector service.type NodePort
If connector is using service.type=NodePort, get the node list and find
the first external address. If no external address is available,
fallback to using the internal IP but warn the user. The internal IP is
usable iff the client has a route to that network.

May be worth adding an option to prefer the internal IP which will also
disable the warning.