improve: use string or file for access key
Previously the secret provider allowed us to read the value from env,
file or literal string, but we had to prefix the value with `env:` or
`file:`.

With this new type we can read the value from all the same places,
without needing that prefix.

This is technically a breaking change, but should be handled by the helm
chart.