fix: filtering of destinations in list cmd
This commit fixes the comparison of grant subject to destination name by using an exact
match or prefix match with trailing dot.

Also refactors to remove the extra iteration. We can build the list of resource names during the
first iteration, each time we find the resource does not yet exist in the map.