fix(helm): create clusterroles based on values not lookup
Using lookup is not stable since the outcome is dependent on the state
of the system. It becomes essentially a toggle as the user
installs/upgrades the connector.

Example:
1. No `logs` cluster role so Helm includes and install it
1. Lookup finds `logs` so Helm omits and removes it
1. Go to 1.

The original intent with the look up is to prevent overriding any
existing roles with the same name.

Instead of using lookup, the chart now contains values to include
or omit specific roles. If a role already exists, Helm will refuse
to overwrite it and the user can disable the role with
`connector.defaultRoles.roles.<role>=false`.

If the user instead wants to disable creating default roles entirely,
they can set `connector.defaultRoles.create=false`.