fix: reduce request transaction lifetime
PostgreSQL row and table locks are held until the end of the
transaction.

Previously we were using a single transaction for the entire lifetime of
the request. One of the first things we did with that transaction was
update the access key and user record associated with the request.

That update would block any other requests from that user until the
first request finished. For short lived requests that was fine, but
longer requests (anything making HTTP requests to an IDP, or blocking
and waiting for updates) would prevent the user from making concurrent
requests.

This commit fixes the problem by committing the "middleware" transaction
first, then starting a new transaction for the request handler.
The middleware transaction should be short lived, because does only a
few database operations and returns. It doesn't block or make external
requests.