improve: combine two similar functions
These two functions were performing very similar operations, one for routes requiring
authentication, and one for optional authentication.

This PR combines them to remove the duplication.  It also extracts a routeSettings struct
so that the settings can be passed around without needing the generic parameters on route.

Also renames a couple of the settings to be more obvious.