improve: use sql for delete destination
Previously we were returning ErrNotFound if the ID did not match any destination,
but we don't do that for any other entity, and no tests appear to expect that behaviour.

Removing that for now for consistency, but we can re-introduce it by checking the
result for the number of updated rows if we want to restore it.