feat(connector): create destination as soon as possible
Update the connector to create itself as quickly as possible instead of
waiting for a valid connection. This is mainly due to LoadBalancer
potentially taking a significant amount of time to initialize. Creating
the destination quickly allows the user to continue without waiting.

Update UI and CLI to output status == "Pending" when the destination is
disconnected and does not have a connection URL.