Move forgot domains query to organizations
The files in our data package are roughly organized by database table.
The forgot domains query is a join between users and organizations, but
it seems like conceptually it is mostly about organizations, so I'ved
moved it to that file.

Also use scanRows and improve asserts.

Also move formatting of the time to the template, so that it's easier to
test the internals, and because a database query function is not the
right place to format a time.

Also remove a function with a boolearn argument.