improve: buffer db initialize in tests
To reduce overall test runtime. By buffering up to 10 calls to NewDB we reduce
the test runtime by 2-3 seconds, about 15%.

The problem is that we can no longer use the package level SymmetricKey because it needs to be
set for NewDB, which now runs in parallel.

In the future we should look to make it not a package level static variable.