fix: change grant channel names
Postgres channel names are limited to 63 bytes. The name of a destination can be longer, and we need space
to store the name of the postgres schema and other identifiers in the channel name as well.

Adds postgres schema to the channel name to prevent tests from overlapping each other.

Also removes destination name from the channel name and uses a single channel per organization.