improve: don't query with empty slice of IDs
listing ByIDs with an empty or nil slice returns all the providers. Previously the required selector
would generate something like 'where ID in ()', which I guess matches no rows.

Now we need to skip the query when the input is an empty slice

Also remove some now unused selectors, and pass t to the setup function so that errors end up
associated with the right test case.