improve: split up api.request function
In preparation for adding the context.Context argument, reduce the number of arguments to
request function by splitting it into two: one that builds the http.Request, and the other
that performs the request and decodes the response body.

This is slightly more verbose, but we still have the http-method-specific helpers to handle
the parts that are common to each method.