Fix response body for unknown paths
This problem is most obvious when a new client is making requests to an old server that
does not yet have the new API path.

Previously the response would be an HTML page from the UI.

With this fix it should be a JSON API error response, which is much easier to debug.