Fix update destination
Start with the existing data so that we don't lose values for fields that
are not updated. We have 3 non-updated fields now: CreatedAt,
LastSeenAt, and Kind.

Also remove the extra logic in the data layer, we no longer need this
now that we're handling it properly in the API.  This allows the
update operation for destination to look like all the other update
operations.

Also add tests for the API endpoint.