fix(ui): trim leading and trailing whitespace in group names
leading whitespaces are particularly problematic because it makes it
possible to create group names that look empty, e.g. " ". a side effect
is group names with " " (one whitespace) and "  " (multple consecutive
whitespaces) will cause a 409 Conflict on creation since the server trims
inputs.

non-leading or non-trailing whitespaces are left as is so it's possible
to have repeats in t between other characters