Use the transaction context in queries
The stdlib only uses the context passed to DB.Begin to cancel
commit.

We don't have a use case yet for a separate query context, so let's use
the context passed to Begin in all the queries as well.

This ensures that if the context is cancelled it will stop any active
queries (instead of allowign the queries and failing at the end on
commit).