Restore context cancellation to background jobs
In an earlier commit we lost cancellation when the artificial transaction was removed.
We can no longer set a context on a data.DB (because DB is not a request
scoped thing it should not have a context associated with it).

And data.Transaction can only accepts an sql.Tx (not an sql.DB), so we
can no longer artificially construct a data.Transaction with a DB
reference like we did with gorm.

This commit creates a new transaction for each run of a background job,
and commits that transaction when it ends (or does a rollback if the job
fails).

This required a bit of refactoring so that the logging was outside of
the function that can return errors.