Do not encrypt empty values
This is a small optimization I noticed while debugging a test failure.
We don't need an expensive encryption process if the value is an empty
string.

This change is backwards compatible because we can still read
encrypted payloads of empty values from the field, but on the next save
it should save it as the empty string, instead of an encrypted payload
that contains an empty string.