Create SSHLoginName in the migration for all users
This will work for now, but there are some unfortunate limitations.

In two places (one in the migration, and again in the test) we have to cast the transaction to its concrete
type to be able to use it.

The migration is also forced to call production code. This can break the migration in the future if the
user table changes.

It should be fine as long as people upgrade their server before we make any substantial changes to the user
model.