Make destination.UniqueID an optional string in the DB
We previously made it an optional string in the API, but because we have
a unique constraint on this field, only a single destination was able to
omit the uniqueID.

This fixes the issue by saving a null in the field. A null omits the row
from the index, which prevents the conflict on uniqueID.

We currently have on destination with an empty uniqueID in dev. That
destination will either get updated on the next deploy with this change,
or we can delete it.