Fix ssh keys test
Previously each test case in this test depeneded on the previous test
case running. Now that we have a way to delete keys, we can clean up
the keys from the previous test case. This allows each test case to
be run independently, and should also make test failures easier to
debug.