Fix a bug in config loading
A slice of structs was causing the location of the configuration value
to be incorrect. There were multiple "exits" which removes the
INFRA_SERVER case from the location. This resulting in any configuration
after a slice of structs to not be loaded.

This bug impacted loading of env variables and command line flags, but
only when some values were previously loaded from a config file.

Using only a config file, or only env vars would have previously worked.

The test added in this commit was failing before the fix.