Fix a bug in SSH known keys
For some reason that I do not yet understand, when this
UserKnownHostsFile is set with a ~/.ssh/infra/known_hosts path, ssh does
not use that file in some cases.

The docs https://man.openbsd.org/ssh_config#UserKnownHostsFile say it
should be supported, and I believe I saw it working previously, but in
the test setup I could not get it to work.

Running ssh with -vvv does not say anything about why that file was
being ignored, and it never tried to read the file.

My best guess is that it has to do with setting HOME to /tmp.

We can use absoluate paths here, because we're already using an absolute
path for the identity file, so I'm going to do that.